+++ webviz_subsurface/_providers/ensemble_surface_provider/ensemble_surface_provider_factory.py
@@ -61,13 +61,19 @@ class EnsembleSurfaceProviderFactory(WebvizFactory):
         return factory
 
     def create_from_ensemble_surface_files(
-        self, ens_path: str, attribute_filter: List[str] = None
+        self,
+        ens_path: str,
+        rel_surface_folder: str = "share/results/maps",
+        attribute_filter: List[str] = None,
     ) -> EnsembleSurfaceProvider:
         timer = PerfTimer()
         string_to_hash = (
-            f"{ens_path}"
+            f"{ens_path}_{rel_surface_folder}"
             if attribute_filter is None
-            else f"{ens_path}_{'_'.join([str(attr) for attr in attribute_filter])}"
+            else (
+                f"{ens_path}_{rel_surface_folder}_"
+                f"{'_'.join([str(attr) for attr in attribute_filter])}"
+            )
         )
         storage_key = f"ens__{_make_hash_string(string_to_hash)}"
         provider = ProviderImplFile.from_backing_store(self._storage_dir, storage_key)
@@ -86,7 +92,7 @@ class EnsembleSurfaceProviderFactory(WebvizFactory):
 
         timer.lap_s()
         sim_surface_files = discover_per_realization_surface_files(
-            ens_path, attribute_filter
+            ens_path, rel_surface_folder, attribute_filter
         )
         obs_surface_files = discover_observed_surface_files(ens_path, attribute_filter)
         et_discover_s = timer.lap_s()